Update Optifine 1.8.9 HD U L5 to 1.8.9 HD U M5

+++ net/minecraft/client/particle/EntityDropParticleFX.java
@@ -128,7 +128,7 @@ public class EntityDropParticleFX extends EntityFX
 
             if (iblockstate.getBlock() instanceof BlockLiquid)
             {
-                d0 = (double)BlockLiquid.getLiquidHeightPercent(((Integer)iblockstate.getValue(BlockLiquid.LEVEL)).intValue());
+                d0 = (double)BlockLiquid.getLiquidHeightPercent(iblockstate.getValue(BlockLiquid.LEVEL));
             }
 
             double d1 = (double)(MathHelper.floor_double(this.posY) + 1) - d0;
