+++ net/minecraft/item/crafting/RecipeFireworks.java
@@ -129,7 +129,7 @@ public class RecipeFireworks implements IRecipe
                     {
                         if (itemstack2.getItem() == Items.dye)
                         {
-                            list.add(Integer.valueOf(ItemDye.dyeColors[itemstack2.getMetadata() & 15]));
+                            list.add(ItemDye.dyeColors[itemstack2.getMetadata() & 15]);
                         }
                         else if (itemstack2.getItem() == Items.glowstone_dust)
                         {
@@ -162,7 +162,7 @@ public class RecipeFireworks implements IRecipe
 
                 for (int l2 = 0; l2 < aint1.length; ++l2)
                 {
-                    aint1[l2] = ((Integer)list.get(l2)).intValue();
+                    aint1[l2] = list.get(l2);
                 }
 
                 nbttagcompound2.setIntArray("Colors", aint1);
@@ -183,7 +183,7 @@ public class RecipeFireworks implements IRecipe
                     {
                         if (itemstack1.getItem() == Items.dye)
                         {
-                            list1.add(Integer.valueOf(ItemDye.dyeColors[itemstack1.getMetadata() & 15]));
+                            list1.add(ItemDye.dyeColors[itemstack1.getMetadata() & 15]);
                         }
                         else if (itemstack1.getItem() == Items.firework_charge)
                         {
@@ -197,7 +197,7 @@ public class RecipeFireworks implements IRecipe
 
                 for (int j2 = 0; j2 < aint.length; ++j2)
                 {
-                    aint[j2] = ((Integer)list1.get(j2)).intValue();
+                    aint[j2] = list1.get(j2);
                 }
 
                 if (this.field_92102_a != null && this.field_92102_a.hasTagCompound())
