+++ net/minecraft/util/MessageSerializer.java
@@ -42,7 +42,7 @@ public class MessageSerializer extends MessageToByteEncoder<Packet>
         else
         {
             PacketBuffer packetbuffer = new PacketBuffer(p_encode_3_);
-            packetbuffer.writeVarIntToBuffer(integer.intValue());
+            packetbuffer.writeVarIntToBuffer(integer);
 
             try
             {
