+++ net/minecraft/world/storage/MapStorage.java
@@ -171,7 +171,7 @@ public class MapStorage
                     {
                         NBTTagShort nbttagshort = (NBTTagShort)nbtbase;
                         short short1 = nbttagshort.getShort();
-                        this.idCounts.put(s, Short.valueOf(short1));
+                        this.idCounts.put(s, short1);
                     }
                 }
             }
@@ -187,22 +187,22 @@ public class MapStorage
      */
     public int getUniqueDataId(String key)
     {
-        Short oshort = (Short)this.idCounts.get(key);
+        Short oshort = this.idCounts.get(key);
 
         if (oshort == null)
         {
-            oshort = Short.valueOf((short)0);
+            oshort = (short)0;
         }
         else
         {
-            oshort = Short.valueOf((short)(oshort.shortValue() + 1));
+            oshort = (short)(oshort + 1);
         }
 
         this.idCounts.put(key, oshort);
 
         if (this.saveHandler == null)
         {
-            return oshort.shortValue();
+            return oshort;
         }
         else
         {
@@ -216,7 +216,7 @@ public class MapStorage
 
                     for (String s : this.idCounts.keySet())
                     {
-                        short short1 = ((Short)this.idCounts.get(s)).shortValue();
+                        short short1 = this.idCounts.get(s);
                         nbttagcompound.setShort(s, short1);
                     }
 
@@ -230,7 +230,7 @@ public class MapStorage
                 exception.printStackTrace();
             }
 
-            return oshort.shortValue();
+            return oshort;
         }
     }
 }
