+++ net/optifine/util/EntityUtils.java
@@ -19,14 +19,14 @@ public class EntityUtils
 
     public static int getEntityIdByClass(Class cls)
     {
-        Integer integer = (Integer)mapIdByClass.get(cls);
-        return integer == null ? -1 : integer.intValue();
+        Integer integer = mapIdByClass.get(cls);
+        return integer == null ? -1 : integer;
     }
 
     public static int getEntityIdByName(String name)
     {
-        Integer integer = (Integer)mapIdByName.get(name);
-        return integer == null ? -1 : integer.intValue();
+        Integer integer = mapIdByName.get(name);
+        return integer == null ? -1 : integer;
     }
 
     public static Class getEntityClassByName(String name)
@@ -62,8 +62,8 @@ public class EntityUtils
                         Config.warn("Duplicate entity name: " + s + ", class1: " + mapClassByName.get(s) + ", class2: " + oclass);
                     }
 
-                    mapIdByClass.put(oclass, Integer.valueOf(i));
-                    mapIdByName.put(s, Integer.valueOf(i));
+                    mapIdByClass.put(oclass, i);
+                    mapIdByName.put(s, i);
                     mapClassByName.put(s, oclass);
                 }
             }
